i3c: i3c-msm-geni: Fix shift out-of-bounds in address slot operations
UBSAN reports a shift out-of-bounds error during boot when
CONFIG_UBSAN_SHIFT is enabled. This is caused by left shift operations
on 32-bit integers with shift values of 32 or more in address slot
manipulation functions. The issue was initially observed in
set_new_addr_slot(), but similar patterns exist in clear_new_addr_slot()
and is_new_addr_slot_set().

Update all affected functions to cast the constant 1 to uint64_t before
performing the shift. This ensures the operation is safely executed
within a 64-bit context, avoiding undefined behavior and preventing
boot failures when UBSAN is active.

+++ drivers/i3c/master/i3c-master-msm-geni.c
@@ -1737,7 +1737,7 @@ static void set_new_addr_slot(unsigned long *addrslot, u8 addr)
 		return;
 
 	ptr = addrslot + (addr / BITS_PER_LONG);
-	*ptr |= 1 << (addr % BITS_PER_LONG);
+	*ptr |= (uint64_t)1 << (addr % BITS_PER_LONG);
 }
 
 static void clear_new_addr_slot(unsigned long *addrslot, u8 addr)
@@ -1748,7 +1748,7 @@ static void clear_new_addr_slot(unsigned long *addrslot, u8 addr)
 		return;
 
 	ptr = addrslot + (addr / BITS_PER_LONG);
-	*ptr &= ~(1 << (addr % BITS_PER_LONG));
+	*ptr &= ~((uint64_t)1 << (addr % BITS_PER_LONG));
 }
 
 static bool is_new_addr_slot_set(unsigned long *addrslot, u8 addr)
@@ -1759,7 +1759,7 @@ static bool is_new_addr_slot_set(unsigned long *addrslot, u8 addr)
 		return false;
 
 	ptr = addrslot + (addr / BITS_PER_LONG);
-	return ((*ptr & (1 << (addr % BITS_PER_LONG))) != 0);
+	return ((*ptr & ((uint64_t)1 << (addr % BITS_PER_LONG))) != 0);
 }
 
 static int qcom_geni_i3c_conf(struct geni_i3c_dev *gi3c, enum i3c_bus_phase bus_phase)
